First pass at smooth facetting + quad meshing
Right now, everything is happening in the evaluator class for Camal. (spaghetti code).
(SmoothMeshEval)
The test can be run with cgm or with the new evaluator. the results are very similar for a simple model, but not exactly the same. 
 to run a simple test:
  ./SmoothIceSheet_test PB.h5m out2.vtk -s 100
PB.h5m file is "geometrized", it has sets for surface, boundary edge, and one vertex

Right now, CGM is linked in, but it is used only if the option -cgm is set.

The correct version would make the smooth evaluator part of MOAB, among other things.
 
This is a work in progress; 
LD_LIBRARY_PATH needs to point to Camal installation