Second pass on smooth facetting and quad meshing the input file is now a triangular mesh for which the geometry sets, sense tags are correctly defined. An example file is shell.h5m, which has 2 surfaces, 7 geometry edges, and 6 vertices. One edge is common to both faces
The code is still a spaghetti, the smooth evaluators are still here, instead
of being defined somewhere in MOAB, in a utility library.
The meshers are also called from the evaluators, this part needs to stay in 
MeshKit

Currently, camal meshers are used, although the edge mesher is not really 
needed, it is a very simple uniform mesher.

CMLPaver will be also replaced.

-cgm option was removed, the current implementation is pretty stable.

-d option will generate vtk *.Point3D files for debugging, with control points
and with some sample evaluations in a grid (not to be used for a large model)

CreateShell.cpp is used to generate the sample file shell.h5m
(it generates nodes, edges, triangles, almost manually, and the correct sets,
parent-child relations, and sense tags on edges)

More work is also needed to evenify the boundary mesh count. If one mesh 
count in a face is odd, all the edges are forced to be even