algs/make_watertight/Makefile.am: correcting old-style definitions of cpp flags algs/Qslim/(several): adding additional includes to make compatible with gcc4.4 algs/Makefile.am: adding targets for a bunch of tests iGeom.hh: updating for latest iGeom utils/Makefile.am: correcting case on iGeom make vars configure.ac: use a better config check for FC; correct a few undefined or mis-defined   config variables (HAVE_CAMAL_EDGE, HAVE_CAMAL_TRIDELAUNAY); alphabetize and add to   Makefile list generated QuadMesh/Makefile: using my own location for msq Makefile.am: inserted check for src dir (new directory for new design); correcting case   in iGeom make vars
And, what caused me to do this in the first place: adding a slew of new class declarations
and some definitions for new design, along with doxygen generation input.  NOTE: the config
turns this off by default, since it doesn't compile yet.  I just need to get this into the
repo to back it up, so I'm committing it now (yes, I know if I were using git I wouldn't have
to do this...)