Make --enable-debug imply --enable-silent-rules
The default can always be overrriden with --(en|dis)able-silent-rules
at configure time or at build time with "make V=1" or "make V=0".

My thinking is that we should have silent rules for developers so that 
compiler warnings and other useful output isn't burried in copious make
output, and developers can always override it with "make V=1".  But for
users, who are presumably compiling w/out --enable-debug, the build 
output is irrelevant except when the build fails, in which case they
probably want to see the build line that failed.