populate_mesh is changed in such a way that if no model entities are found after parsing the geometry entities and geometry mesh sets, one model entity with a moab set is created, comprising the highest dimension elements from the mesh database.
With this change, loading a simple mesh file with no sets in it will result 
in at least one model ent created. Qslim test will then not need to create 
an artificial model ent, so the test is simplified.

An example python script is also shown to illustrate how to call Qslim Mesher 
using python in Meshkit.