Change a setting in make_watertight
With a current version of MOAB/DagMC, having this boolean set
to true causes DagMC queiries to be made when DagMC has not
been fully initialized, causing segfaults.  I do not know how
important it is to have this check performed; this commit is
made to allow the tool to at least function.