Interval Assignment mapping test progress
Tim, I need help with this part:
  moab::Range curves;
  this_surf->boundary(1, curves); // Moab? Find the right function call
  CHECK_EQUAL(4, (int)curves.size());
  // assume curves are ordered 0 to 3 contiguously around the surface
  //   or perhaps 0, 1 are opposite, and 2, 3 are opposite?
  // for (MeshKit::MEntVector::iterator vit = curves.begin(); ...

  MeshKit::ModelEnt* me_curves[4] = {0,0,0,0};
  // convert moab entity handles to ModelEnt* and place in me_curves... ask Tim how

I assume me_curves[0] is opposite to me_curves[2], same with 1 and 3.
I.e. they are ordered around the loop 0,1,2,3

Also, I expect the mapping_test to currently fail because the mesh entities are null. But I also get this error and could use help with it:
Running mapping_test ...
terminate called after throwing an instance of 'MeshKit::Error'
  what():  MKCore.cpp, line 174: Failure to create MKModelEnt tag in iGeom.
  mapping_test: TERMINATED (signal 6)
FAIL: test_intassign

This is happening the *second* time I try to create an interval assignment entity, and is not specific to the mapping test. For example, it also occurs for test_many_pairs.

--
The many_pairs test is failing to produce good values, and I don't yet know why.
Don't worry about that, I'll fix soon.

--