test_ebmesh is including directly this: #include "CGMApp.hpp" and is also using  CGMApp::instance()
this is defined in libcubit_geom.so library. 

for ubuntu 12.04, the dependencies to the external libraries need to be explicit, so this is why it is added 
to the LDADD variable for test_ebmesh ( -lcubit_geom ) 
(it is needed only for OCC based builds, as cubit based builds already have this explicit dependency)

I think that we should remove this dependency, it should go through iGeom API only;
we should not add a CGM dependency in meshkit code.

But until we fix test_ebmesh, this solution should be good enough.
And it should not break any build, either it based on cgm 12.2 or CGM trunk. 
We will see tonight ...