Compute num layers from sizing fnctn in TFIMapping on cylinder.
Previously this required that the number of intervals be explicitly set on the
sizing function.  Now if the size is set and not the number of intervals, the
number of intervals will be computed based on the size.  This applies only
if there is a cylinder without a linking edge.