Add some initial work towards implementing 2-dimensional advancing front.
This initial work is subject to change.  So far it has some pieces that will be needed to define a rule and an implementation of a local transformation that can transfer points between three dimensions and two dimensions.  Tests for the implementation of the free zone and the local transformation are included.