Add more work towards a 2-D advancing front algorithm.
The primary contribution of this commit is AF2Front, which is an object that manages the "front" of the advancing front algorithm.

An AF2LocalTransformMaker concept is introduced to manage building local transformations between the surface and some 2-dimensional space.  Thus the front itself does not have to be responsible for the transformation.

Some modifications were required, particularly to AF2Edge3D, to remove const qualifiers that don't work well with the way the front tracks quality levels.

Tests of AF2Front and an AF2LocalTransformMaker are included.