Fix a bug in AF2FreeZone and add a regression test for it.
The bug had to do with intersection detection for a test edge that had an endpoint matching a free zone vertex when the angle between the edge and the free zone boundary was moderately small.

This commit also removes an unintentional include of iostream from AF2Front.

This commit also moves a variable to global scope in the frontObj2D test to help ensure that a different pointer variable at global scope remains a valid pointer.