Implement AF2DfltTriangleMeshOp to integrate 2-d advancing front in MeshKit.
Some small modifications to AF2Algorithm and AF2Point3D were required to streamline the storage and commitment of MOAB vertices.

The new MeshOp is always registered with MeshKit, but isn't (yet) the default MeshOp for 2-d.

A test is added that tests using the MeshOp on a variety of flat surfaces with and without holes, as well as a sphere and a piece of a torus.